requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Jinja2==2.10.3
numpy==1.17.4
pymongo==3.9.0
mysql-connector-python==8.0.31
selenium
pillow
requests
dnspython
